convert toggle to typescript
