--- src/components/toggle/index.js
@@ -1,4 +1,0 @@
-export {
-  EuiToggle,
-  TYPES as TOGGLE_TYPES,
-} from './toggle';
