+++ src/components/toggle/index.ts
@@ -1,0 +1,1 @@
+export { EuiToggle, ToggleType, TYPES as TOGGLE_TYPES } from './toggle';
