+++ src/test/index.d.ts
@@ -1,4 +1,5 @@
 import { ShallowWrapper, ReactWrapper } from 'enzyme';
+export { requiredProps } from './required_props';
 
 declare module '@elastic/eui' {
   export function findTestSubject<T extends ShallowWrapper | ReactWrapper> (
